refactor, ts-jest, tree shake esm, cjs build, typedef
