+++ src/types.ts
@@ -1,0 +1,7 @@
+export interface InitializeKeypairOptions {
+  envFileName?: string;
+  envVariableName?: string;
+  airdropAmount?: number | null;
+  minimumBalance?: number;
+  keypairPath?: string;
+}
