+++ tsconfig.json
@@ -1,15 +1,20 @@
-// See https://aka.ms/tsconfig
 {
   "compilerOptions": {
-    "outDir": "dist",
-    "target": "es2022",
-    "moduleResolution": "Bundler",
-    "module": "ES2022",
+    "target": "ES2018",
+    "lib": ["ES2018", "DOM"],
+    "moduleResolution": "node",
+    "outDir": "./dist/esm",
+    "rootDir": "./src",
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "module": "ESNext",
+    "isolatedModules": true
   },
-  "include": ["src"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
 }
