+++ tsconfig.test.json
@@ -1,0 +1,19 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "types": ["node", "jest"],
+    "module": "ESNext",
+    "moduleResolution": "node",
+    "target": "ES2022",
+    "noEmit": true,
+    "allowImportingTsExtensions": true
+  },
+  "include": ["src/**/*", "tests/**/*"],
+  "exclude": [
+    "node_modules",
+    "dist",
+    "tests/bankrun_test",
+    "coverage/**",
+    "**/*.d.ts",
+  ]
+}
